Move core.Tensor into framework/ops.py.
PiperOrigin-RevId: 165034131
